\name{qwraps-package}
\alias{qwraps-package}
\alias{qwraps}
\docType{package}
\title{
  Quick Wraps
}
\description{
  Wrapper functions for producing common R base plots not explicitly defined in
  the ggplot2 package such as ACF and ROC plots.  Other wrapper functions for
  formating results and extracting parameter estimates to be used in
  Sweaved/Knited reports.
}
\details{
\tabular{ll}{
Package: \tab qwraps\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-12\cr
License: \tab GPL >=2\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
What license is it under?}
