\name{simpleCap}
\alias{simpleCap}
\title{Simple Capitalization}
\usage{
  simpleCap(x)
}
\arguments{
  \item{x}{string to edit}
}
\value{
  The requested reformatted string.
}
\description{
  Remove a "." from a string and capitalize the first
  letter of every word.
}
\note{
  Slight modification of the function noted in the
  exampleof the chartr base help file.
}
\examples{
s <- "this.is.a.string"
simpleCap(s)
}
\author{
  Peter DeWitt
}
\references{
  %% ~put references to the literature/web site here ~
}
\seealso{
  %% ~~objects to See Also as \code{\link{help}}, ~~~
}
\keyword{~kwd1}
\keyword{~kwd2}

